requirements.txt
alembic==1.14.0
aniso8601==9.0.1
bcrypt==4.2.1
blinker==1.9.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
cloudinary==1.41.0
Flask==3.1.0
Flask-Bcrypt==1.0.1
Flask-Cors==5.0.0
Flask-JWT-Extended==4.7.0
Flask-Mail==0.10.0
Flask-Migrate==4.0.7
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.6
MarkupSafe==3.0.2
psycopg2-binary==2.9.9
PyJWT==2.10.0
pyotp==2.9.0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
setuptools==70.3.0
six==1.16.0
SQLAlchemy==2.0.29
sqlalchemy-serializer==1.4.22
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.1.3
